utils: add get_pid_type_from_ref
