+++ inspirehep/utils/helpers.py
@@ -28,6 +28,10 @@ from contextlib import closing
 
 import requests
 
+from six.moves.urllib.parse import urlsplit
+
+from inspirehep.modules.pidstore.utils import get_pid_type_from_endpoint
+
 
 def download_file(url, output_file=None, chunk_size=1024):
     """Download a file to specified location."""
@@ -90,3 +94,15 @@ def get_recid_from_url(reference):
     except (ValueError, AttributeError):
         res = False
     return res
+
+
+def get_pid_type_from_ref(reference):
+    """Return the ``pid_type`` corresponding to a record URL.
+
+    The reference name corresponds to the ``endpoint`` in all cases.
+    This implementation exploits this by falling back to
+    ``get_pid_type_from_endpoint``.
+    """
+    endpoint = urlsplit(reference).path.split('/')[-2]
+
+    return get_pid_type_from_endpoint(endpoint)
