+++ tests/unit/utils/test_utils_helpers.py
@@ -21,7 +21,7 @@
 # or submit itself to any jurisdiction.
 
 from inspirehep.dojson.utils import force_force_list
-from inspirehep.utils.helpers import get_recid_from_url
+from inspirehep.utils.helpers import get_recid_from_url, get_pid_type_from_ref
 
 
 def test_force_force_list_returns_empty_list_on_none():
@@ -59,3 +59,10 @@ def test_get_recid_from_url_returns_false_on_ref_malformed():
 
 def test_get_recid_from_url():
     assert get_recid_from_url('http://localhost:5000/api/literature/111') is 111
+
+
+def test_get_pid_type_from_ref():
+    expected = 'lit'
+    result = get_pid_type_from_ref('http://localhost:5000/api/literature/111')
+
+    assert expected == result
